requirements.txt
PyGithub==1.57
python-dotenv==0.21.0
pandas==1.5.2